+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [unreleased]
+* use `wp-content/temp-backup` and not `wp-content/upgrade/temp-backup` as `WP_Upgrader::unpack_package` deletes contents of `wp-content-upgrade` at each update
 
 #### 3.2.1 / 2022-09-23
 * bump auto-deactivation check for WP version